backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
sse-starlette>=1.6.0
anthropic>=0.39.0
ddgs>=6.0.0
yfinance>=0.2.31
python-docx>=1.1.0
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
pydantic>=2.5.0
pandas>=2.1.0
openpyxl>=3.1.0
python-dotenv>=1.0.0
